+++ extensions/vscode/src/api/index.ts
@@ -6,7 +6,7 @@ export * from "./types/accounts";
 export * from "./types/credentials";
 export * from "./types/configurations";
 export * from "./types/connect";
-export * from "./types/deployments";
+export * from "./types/contentRecords";
 export * from "./types/events";
 export * from "./types/files";
 export * from "./types/schema";